ROS log data warning

+++ docker/utils/save_latest_ros_logs.sh
@@ -35,18 +35,73 @@ function copy_to_host
     rm -rf ${ros_log_dir}/$dir/ros_log_$timestamp/logs_temp
 }
 
+container_name=$(docker ps | awk '{if(NR>1) print $NF}')
+
 echo -e "${NC}${normal}You are about to save latest ros logs ${normal}${NC}"
-echo -e "${RED}${bold}WARNING! This closes all running docker containers. Do you wish to continue? (y/n) ${normal}${NC}"
+echo -e "${YELLOW}${bold}WARNING! This closes all running docker containers. Do you wish to continue? (Y/n) ${normal}${NC}"
 read prompt
 
-if [[ $prompt == "" || $prompt == "n" || $prompt == "N" || $prompt == "no" || $prompt == "No" || $prompt == "NO" ]]; then
+if [[ $prompt == "n" || $prompt == "N" || $prompt == "no" || $prompt == "No" || $prompt == "NO" ]]; then
     exit 1
 fi
 
-echo -e "${NC} ${normal}Please add a note for logging with reasons... ${normal}${NC}"
+echo -e "${NC}${normal}Please add a note for logging with reasons... ${normal}${NC}"
 read notes_from_user
 
-container_name=$(docker ps | awk '{if(NR>1) print $NF}')
+if [ "$(docker exec ${container_name} bash -c 'ls /usr/local/bin/customer.key')" ]; then
+    save_log_msg_config_file="/home/$USER/.shadow_save_log_app/save_sr_log_msg_config.cfg"
+    tmp_save_log_msg_config_file="/home/$USER/.shadow_save_log_app/tmp_save_sr_log_msg_config.cfg"
+
+    if [ -f $save_log_msg_config_file ]; then
+        # check if the file contains something we don't want
+        if egrep -q -v '^#|^[^ ]*=[^;&]*' "$save_log_msg_config_file"; then
+          echo "Config file is unclean, cleaning it..." >&2
+          # filter the original to a tmp file
+          egrep '^#|^[^ ]*=[^;&]*'  "$configfile" > "$tmp_save_log_msg_config_file"
+          mv $tmp_save_log_msg_config_file $save_log_msg_config_file
+        fi
+    else
+        touch $save_log_msg_config_file
+        echo 'do_not_show_upload_log_message="false"' >> $save_log_msg_config_file
+        echo 'upload_sr_log_messages="true"' >> $save_log_msg_config_file
+    fi
+
+    source $save_log_msg_config_file
+
+    if [ ! $do_not_show_upload_log_message == "true" ]; then
+        counter=0
+        while ! [[ $upload_to_server == "Y" || $upload_to_server == "y" || $upload_to_server == "yes" || $upload_to_server == "YES" || $upload_to_server == "N" || $upload_to_server == "n" || $upload_to_server == "no" || $upload_to_server == "NO" ]]; do
+            if [ $counter -gt 4 ]; then
+                echo -e "${RED}Too many invalid inputs. Exiting the program...${normal}${NC}"
+                sleep 5
+                exit 1
+            fi
+            echo -e "${YELLOW}We are going to upload logs to Shadow servers so we can diagnose problems. Do you want to do this? (Y/n) ${normal}${NC}"
+            read upload_to_server
+            if ! [[ $upload_to_server == "Y" || $upload_to_server == "y" || $upload_to_server == "yes" || $upload_to_server == "YES" || $upload_to_server == "N" || $upload_to_server == "n" || $upload_to_server == "no" || $upload_to_server == "NO" ]]; then
+                echo "Please type 'Y' or 'n'"
+            fi
+            let counter+=1
+        done
+        echo -e "${YELLOW}If you don't want to see the previous message again, type 'Y'. Otherwise, type 'n' (Y/n) ${normal}${NC}"
+        read dont_show_upload_log_message
+
+        if [[ $dont_show_upload_log_message == "Y" || $dont_show_upload_log_message == "Yes" || $dont_show_upload_log_message == "y" || $dont_show_upload_log_message == "YES" || $show_upload_log_message == "NO" ]]; then
+            sed -i 's/\(do_not_show_upload_log_message *= *\).*/\1"true"/' $save_log_msg_config_file
+        else
+            sed -i 's/\(do_not_show_upload_log_message *= *\).*/\1"false"/' $save_log_msg_config_file
+        fi
+
+        if [[ $upload_to_server == "N" || $upload_to_server == "No" || $upload_to_server == "n" || $upload_to_server == "no" || $upload_to_server == "NO" ]]; then
+            sed -i 's/\(upload_sr_log_messages *= *\).*/\1"false"/' $save_log_msg_config_file
+            upload_sr_log_messages="false"
+            echo -e "${YELLOW}Your logs will be stored locally and not uploaded to Shadow servers! ${normal}${NC}"
+        else
+            sed -i 's/\(upload_sr_log_messages *= *\).*/\1"true"/' $save_log_msg_config_file
+            upload_sr_log_messages="true"
+        fi
+    fi
+fi
 
 if [ ! -z "$container_name" ]; then
         container_array=($container_name)
@@ -90,15 +145,16 @@ if [ ! -z "$container_name" ]; then
             if [ ${customerkey} ]; then
              # check if the folder is empty.
                 if [ ! -z "$(docker exec ${container_name} bash -c 'find /home/user/logs_temp -maxdepth 0 -type d 2>/dev/null')" ]; then
-                    echo -e "${RED}${bold}There are previous logs that havent been sent yet. Would you like to send them now? Type 'y' to send or 'n' to ignore and overwrite them ${normal}${NC}"
+                    echo -e "${YELLOW}${bold}There are previous logs that havent been sent yet. Would you like to send them now? Type 'Y' to send or 'n' to ignore and overwrite them ${normal}${NC}"
                     read old_logs
-                    if [[ $old_logs == "y" || $old_logs == "Y" || $old_logs == "yes" ]]; then
-                        echo "Uploading to AWS - Please wait..."
+                    if [[ $old_logs == "y" || $old_logs == "Y" || $old_logs == "yes" || $old_logs == "Yes" || $old_logs == "YES" ]]; then
+                        echo "Uploading to Shadow servers - Please wait..."
                         upload_command=$(docker exec $current_container_name bash -c "source /usr/local/bin/shadow_upload.sh ${customerkey} /home/user/logs_temp $timestamp" || true) 
                         if [[ $upload_command == "ok" ]]; then
-                            echo -e "${GREEN} Previous logs Uploaded to AWS for $current_container_name! ${NC}"
+                            echo -e "${GREEN} Previous logs Uploaded to Shadow servers for $current_container_name! ${NC}"
                         else
-                            echo -e "${RED}${bold} Failed to upload previous logs to AWS for $current_container_name! Check your internet connection and try again. Exiting... ${normal}${NC}"
+                            echo -e "${RED}${bold} Failed to upload previous logs to Shadow servers for $current_container_name! Check your internet connection and try again. Exiting... ${normal}${NC}"
+                            sleep 5
                             exit 1
                         fi
 		                sleep 1
@@ -109,16 +165,20 @@ if [ ! -z "$container_name" ]; then
                 # copy new logs to temp folder
                 copy_logs
                 copy_to_host
-                echo "Uploading to AWS - Please wait..."
-                upload_command=$(docker exec $current_container_name bash -c "source /usr/local/bin/shadow_upload.sh ${customerkey} /home/user/logs_temp $timestamp" || true)
-                if [[ $upload_command == "ok" ]]; then
-                    # delete temp folder
-                    docker exec $current_container_name bash -c "rm -rf /home/user/logs_temp"
-                    echo -e "${GREEN} Latest Logs Saved and Uploaded to AWS for $current_container_name! ${NC}"
-		        else
-                    echo -e "${RED}${bold} Failed to upload logs to AWS for $current_container_name! Check your internet connection and try again.${normal}${NC}"
+                if [ $upload_sr_log_messages == "true" ]; then
+                    echo "Uploading to Shadow servers - Please wait..."
+                    upload_command=$(docker exec $current_container_name bash -c "source /usr/local/bin/shadow_upload.sh ${customerkey} /home/user/logs_temp $timestamp" || true)
+                    if [[ $upload_command == "ok" ]]; then
+                        # delete temp folder
+                        docker exec $current_container_name bash -c "rm -rf /home/user/logs_temp"
+                        echo -e "${GREEN} Latest Logs Saved and Uploaded to Shadow servers for $current_container_name! ${NC}"
+                    else
+                        echo -e "${RED}${bold} Failed to upload logs to Shadow servers for $current_container_name! Check your internet connection and try again.${normal}${NC}"
+                        sleep 5
+                        exit 1
+                    fi
+                    sleep 1
                 fi
-                sleep 1
             else
                 copy_logs
                 copy_to_host
